+++ packages/client/src/views/system/systemMonitor/index.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
+import { BsXLg } from 'react-icons/bs';
 
+import { IconButton } from '$components/iconButton';
 import { Log } from '$components/log';
 import { Tab, TabContent, TabPanel, Tabs, TabView } from '$components/tabs';
 import { mainStore } from '$stores/MainStore';
+import { uiStore } from '$stores/UIStore';
 import styles from './index.module.scss';
 import { Utilization } from './Utilization';
 import { LoadedModelsList } from '../../common/backend/LoadedModelsList';
@@ -24,7 +27,12 @@ export const SystemMonitor: React.FC = () => {
           </TabPanel>
         </TabContent>
       </TabView>
-      <div>
+      <div className={styles.right}>
+        <div className={styles.rightHeader}>
+          <IconButton onClick={() => uiStore.toggleSystemMonitor()}>
+            <BsXLg />
+          </IconButton>
+        </div>
         <Utilization />
       </div>
     </div>